strings.f: add new string allocator
The existing string allocation system is a single big use-after-free bug.
r-alloc deallocates memory as soon as the caller returns, making it
entirely unsuitable for functions returning temporary strings, yet it
is used in all of them!

This allocator uses a ring buffer in EWRAM instead.  This will still fail
if there are too many temporary strings in an expression, but at least
it will do so in a much more predictable fashion.

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>