timers.f: starttimer: enable IME
This is funny.  For starters, this change is only required to fix tutorial 1
for Mednafen, which does not implement SWI 2 correctly and waits forever if
IME is not enabled, which according to specs is incorrect and incongruent
with the real thing and other emulators.

On the other hand, starttimer enables the timer interrupt, which is somewhat
pointless without also enabling IME.

On the third hand, the DB API seems to expect users to handle their
interrupt enabling and disabling themselves, which is not very intuitive now
that many everyday functions rely on interrupts and they need to be on
pretty much all the time.

On the fourth hand, enabling IME avoids the effect observed in NO$GBA that
the BIOS Halt returns very quickly (after 200-1000 cycles), keeping the CPU
toasty even though nothing is enabled besides the timer interrupt.

In the end, this change makes tutorial 1 work everywhere, so that's how it
will be for now.

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>