MF: Parser: do not code pops/pushes reaching into the stack frame
When functions exhaust their own stack use, they attempt to pull TOS
into R0, which in the case of functions with a stack frame is the first
local (or an argument).  This value will never be used by the function
except through an explicit load or store, so the only thing that will
ever happen to it is that it is pushed back, or purged on return.  It is
therefore safe to simply not code such pops and pushes.

"lepilog" has to be changed not to assume that the first local has been
popped into R0.

It is also necessary to make sure this only happens in functions with a
stack frame, sp_offset is set to a high value on all other kinds of
functions.

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>