Handle inline code in functions without base pointer correctly.
OMG.  Inlined code does not know about our underwater stack trick, so
we need to re-add skipped pushes or pops to compensate for its behavior.
In order to do that, we need to know how the inlined code affects the
stack, though, so all assembly-language inlined functions have to be
annotated according to how they affect the stack.  (Luckily this can be
done automatically in inlined TIN words.)

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>