DBC: Allocate maximum size for string variables.
Allocating one word (4 bytes) obviously doesn't work; there does not
seem to be a syntax for specifying the maximum size of a string variable,
so we simply allocate the theoretical maximum of 64 words (256 bytes) for
now. This is obviously not ideal, but at least there won't be any random
data corruption anymore.

Fixes issue #20.

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>